+++ packages/components/checkbox/src/checkbox-group.stories.ts
@@ -2,6 +2,7 @@ import '@sl-design-system/button/register.js';
 import '@sl-design-system/button-bar/register.js';
 import '@sl-design-system/form/register.js';
 import '@sl-design-system/infotip/register.js';
+import '@sl-design-system/tooltip/register.js';
 import { type Meta, type StoryObj } from '@storybook/web-components-vite';
 import { type TemplateResult, html } from 'lit';
 import '../register.js';
@@ -58,6 +59,7 @@ export default {
           ${slot?.() ??
           html`
             <sl-checkbox-group
+              aria-label=${label || 'Label'}
               ?disabled=${disabled}
               ?required=${required}
               .label=${label}
@@ -77,6 +79,11 @@ export default {
           <sl-button @click=${onClick}>Report validity</sl-button>
         </sl-button-bar>
       </sl-form>
+      <style>
+        sl-tooltip {
+          position-area: right;
+        }
+      </style>
     `;
   }
 } satisfies Meta<Props>;
@@ -109,7 +116,7 @@ export const Value: Story = {
 export const ImplicitValue: Story = {
   args: {
     slot: () => html`
-      <sl-checkbox-group>
+      <sl-checkbox-group aria-label="Implicit value options">
         <sl-checkbox checked value="0">Option 1</sl-checkbox>
         <sl-checkbox checked value="1">Option 2</sl-checkbox>
         <sl-checkbox value="2">Option 3</sl-checkbox>
@@ -121,7 +128,7 @@ export const ImplicitValue: Story = {
 export const Infotip: Story = {
   args: {
     slot: () => html`
-      <sl-checkbox-group>
+      <sl-checkbox-group aria-label="Options with infotips">
         <sl-checkbox checked value="0"
           >Option 1 with infotip and a very long label that should wrap to the next line. This label
           is way too long to be used in the label for the "more info" button, so we have provided a
@@ -166,6 +173,19 @@ export const NoLabel: Story = {
   }
 };
 
+export const Tooltips: Story = {
+  args: {
+    boxes: () => html`
+      <sl-checkbox id="1" value="1">Option 1</sl-checkbox>
+      <sl-tooltip for="1" type="description">Tooltip for option 1</sl-tooltip>
+      <sl-checkbox id="2" value="2">Option 2</sl-checkbox>
+      <sl-tooltip for="2" type="description">Tooltip for option 2</sl-tooltip>
+      <sl-checkbox id="3" value="3">Option 3</sl-checkbox>
+      <sl-tooltip for="3" type="description">Tooltip for option 3</sl-tooltip>
+    `
+  }
+};
+
 export const CustomValidity: Story = {
   args: {
     hint: 'This story has both builtin validation (required) and custom validation. You need to select the middle option to make the field valid. The custom validation is done by listening to the sl-validate event and setting the custom validity on the checkbox group.',
@@ -177,7 +197,7 @@ export const CustomValidity: Story = {
       };
 
       return html`
-        <sl-checkbox-group @sl-validate=${onValidate} required>
+        <sl-checkbox-group @sl-validate=${onValidate} aria-label="Your favorite number" required>
           <sl-checkbox value="1">One</sl-checkbox>
           <sl-checkbox value="2">Two</sl-checkbox>
           <sl-checkbox value="3">Three</sl-checkbox>
@@ -207,7 +227,7 @@ export const CustomAsyncValidity: Story = {
       };
 
       return html`
-        <sl-checkbox-group @sl-validate=${onValidate} required>
+        <sl-checkbox-group @sl-validate=${onValidate} aria-label="Your favorite number" required>
           <sl-checkbox value="1">One</sl-checkbox>
           <sl-checkbox value="2">Two</sl-checkbox>
           <sl-checkbox value="3">Three</sl-checkbox>